+++ trixnity-crypto-core/src/jsMain/kotlin/net/folivo/trixnity/crypto/core/sha256.kt
@@ -7,20 +7,25 @@ import js.typedarrays.toUint8Array
 import kotlinx.coroutines.flow.*
 import net.folivo.trixnity.utils.ByteArrayFlow
 import net.folivo.trixnity.utils.encodeUnpaddedBase64
-import net.folivo.trixnity.utils.toByteArray
-import web.crypto.crypto
+import okio.Buffer
+import okio.HashingSink
+import okio.blackholeSink
 
 actual fun ByteArrayFlow.sha256(): Sha256ByteFlow {
     return if (PlatformUtils.IS_BROWSER) {
-        val crypto = crypto.subtle
         val hash = MutableStateFlow<String?>(null)
         Sha256ByteFlow(
-            flow {// TODO should be streaming!
-                val content = toByteArray()
-                hash.value = Uint8Array(crypto.digest("SHA-256", content.toUint8Array()))
-                    .toByteArray()
-                    .encodeUnpaddedBase64()
-                emit(content)
+            flow {
+                val sink: HashingSink = HashingSink.sha256(blackholeSink())
+                emitAll(
+                    onEach {
+                        val buffer = Buffer().write(it)
+                        sink.write(buffer, buffer.size)
+                    }.onCompletion {
+                        hash.value = sink.hash.toByteArray().encodeUnpaddedBase64()
+                        sink.close()
+                    }
+                )
             },
             hash,
         )